alerts/requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
flower==1.1.0
humanize==4.2.3
idna==3.3
ipython==8.4.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
